RaG_N_ROLL/requirements.txt
faker
snowflake-connector-python[pandas]
pandas
python-dotenv